+++ commonmark/src/main/java/org/commonmark/internal/inline/AutolinkInlineParser.java
@@ -3,9 +3,9 @@ package org.commonmark.internal.inline;
 import org.commonmark.node.Link;
 import org.commonmark.node.Text;
 import org.commonmark.parser.SourceLines;
-import org.commonmark.parser.beta.Position;
-import org.commonmark.parser.beta.Scanner;
+import org.commonmark.parser.beta.*;
 
+import java.util.Set;
 import java.util.regex.Pattern;
 
 /**
@@ -45,5 +45,17 @@ public class AutolinkInlineParser implements InlineContentParser {
             }
         }
         return ParsedInline.none();
+    }
+
+    public static class Factory implements InlineContentParserFactory {
+        @Override
+        public Set<Character> getTriggerCharacters() {
+            return Set.of('<');
+        }
+
+        @Override
+        public InlineContentParser create() {
+            return new AutolinkInlineParser();
+        }
     }
 }