+++ commonmark/src/main/java/org/commonmark/internal/inline/BackticksInlineParser.java
@@ -3,9 +3,10 @@ package org.commonmark.internal.inline;
 import org.commonmark.node.Code;
 import org.commonmark.node.Text;
 import org.commonmark.parser.SourceLines;
-import org.commonmark.parser.beta.Position;
-import org.commonmark.parser.beta.Scanner;
+import org.commonmark.parser.beta.*;
 import org.commonmark.text.Characters;
+
+import java.util.Set;
 
 /**
  * Attempt to parse backticks, returning either a backtick code span or a literal sequence of backticks.
@@ -46,5 +47,17 @@ public class BackticksInlineParser implements InlineContentParser {
         SourceLines source = scanner.getSource(start, afterOpening);
         Text text = new Text(source.getContent());
         return ParsedInline.of(text, afterOpening);
+    }
+
+    public static class Factory implements InlineContentParserFactory {
+        @Override
+        public Set<Character> getTriggerCharacters() {
+            return Set.of('`');
+        }
+
+        @Override
+        public InlineContentParser create() {
+            return new BackticksInlineParser();
+        }
     }
 }