+++ commonmark/src/main/java/org/commonmark/parser/beta/InlineParserState.java
@@ -1,13 +1,10 @@
-package org.commonmark.internal.inline;
-
-import org.commonmark.parser.beta.Position;
-import org.commonmark.parser.beta.Scanner;
+package org.commonmark.parser.beta;
 
 public interface InlineParserState {
 
     /**
-     * Return a scanner for the input for the current position (on the character that the inline parser registered
-     * interest for).
+     * Return a scanner for the input for the current position (on the trigger character that the inline parser was
+     * added for).
      * <p>
      * Note that this always returns the same instance, if you want to backtrack you need to use
      * {@link Scanner#position()} and {@link Scanner#setPosition(Position)}.